+++ tests/integration/test_scm_conc.py
@@ -211,8 +211,6 @@ def test_ciceroscm_short_run(tmpdir, test_data_dir):
         + cscm.results["carbon cycle"]["Biosphere carbon flux"].values
         + cscm.results["carbon cycle"]["Ocean carbon flux"].values
     )
-    print(carbon_sum[-5:])
-    print(np.cumsum(cscm.results["emissions"]["CO2"])[-5:])
     # TODO: Why isn't this closer?
     assert np.allclose(
         carbon_sum, np.cumsum(cscm.results["emissions"]["CO2"]), rtol=5e-1